Rename io to fio
