+++ src/fiat/__init__.py
@@ -17,6 +17,6 @@ from osgeo import osr
 osr.UseExceptions()
 
 from .cfg import Configurations
-from .io import open_csv, open_geom, open_grid
+from .fio import open_csv, open_geom, open_grid
 from .models import GeomModel, GridModel
 from .version import __version__
